+++ server/area/restore.php
@@ -264,6 +264,8 @@
                                 $("#sockets").val("<?php echo $sb_status['setting17']; ?>");
                                 $("#cores").val("<?php echo $sb_status['setting18']; ?>");
                                 $("#threads").val("<?php echo $sb_status['setting19']; ?>");
+                                $("#thinprovision").val("<?php echo $sb_status['setting21']; ?>");
+                                $("#passdiscard").val("<?php echo $sb_status['setting22']; ?>");
 								<?php
 								if ( strpos( $sb_status['setting6'], 'fixgrub' ) !== false ) {
 								?>