+++ ChangeLog.md
@@ -3,6 +3,10 @@
  
  [Back To ReadMe](https://github.com/zipurman/oVIRT_Simple_Backup/)
  
+  - 0.6.12 - 2018/04/05
+      - [x] Allow option for thin provisioned (sparse) disks in restore/migrate
+      - [x] When restoring or migrating, if choosing a storage domain that supports discard, then you can optionally pass_discard=true to the VM disk(s)
+      
   - 0.6.11 - 2018/04/05
      - Additional bug fixes with automated backups and errors with fdisk not being found
      - [x] Added extra logic to WebUI if Automated Backup is running